Account for pre-5 behavior
In Puppet4, the Report class requires a kind argument. Also the is_to_s
and should_to_s methods returned an array of strings and a flattened
array as a string. That behavior was changed in PUP-7616 (commit
c14b28f9c427) so that both methods return a string as the name
implies.