(PA-5781) Update stdout and stderr methods for Beaker 5
This commit updates stderr and stdout methods to be called on Result
objects and removes _ in front of variables that are used.

Previously in Beaker, you could use standalone stdout and stderr methods
to access output from remote machines.

These methods were deprecated in 2013 with voxpupuli/beaker@28b2510 and
dropped entirely in voxpupuli/beaker@73a31c7.